adding museum character adding check for voice availability
