+++ OpenAIChat/README.md
@@ -1,2 +1,11 @@
 # OpenAIChat
 
+## Description
+This is a skill template. For more example skills go to furhat https://github.com/FurhatRobotics/
+
+## Usage
+Max number of users is set to: 2
+
+### Requirements:
+API key for OpenAI GPT3 language model. Request it from https://openai.com/api/
+Character Jerry Seinfeld requires Acapela Voice 'WillFromAfar"'
