--- OpenAIChat/README.txt
@@ -1,1 +1,0 @@
-Under development. Coming soon.
