+++ OpenAIChat/src/main/kotlin/furhatos/app/openaichat/flow/chatbot/openai.kt
@@ -1,4 +1,4 @@
-package furhatos.app.openaichat
+package furhatos.app.openaichat.flow.chatbot
 
 import com.theokanning.openai.OpenAiService
 import com.theokanning.openai.completion.CompletionRequest
@@ -12,11 +12,12 @@ class OpenAIChatbot(val description: String, val userName: String, val agentName
 
     val service = OpenAiService(serviceKey)
 
-    var temperature = 0.9
-    var maxTokens = 50
-    var topP = 1.0
-    var frequencyPenalty = 0.0
-    var presencePenalty = 0.6
+    // Read more about these settings: https://beta.openai.com/docs/introduction
+    var temperature = 0.9 // Higher values means the model will take more risks. Try 0.9 for more creative applications, and 0 (argmax sampling) for ones with a well-defined answer.
+    var maxTokens = 50 // Length of output generated. 1 token is on average ~4 characters or 0.75 words for English text
+    var topP = 1.0 // 1.0 is default. An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered.
+    var frequencyPenalty = 0.0 // Number between -2.0 and 2.0. Positive values penalize new tokens based on their existing frequency in the text so far, decreasing the model's likelihood to repeat the same line verbatim.
+    var presencePenalty = 0.6 // Number between -2.0 and 2.0. Positive values penalize new tokens based on whether they appear in the text so far, increasing the model's likelihood to talk about new topics.
 
     fun getNextResponse(): String {
         /** The prompt for the chatbot includes a context of ten "lines" of dialogue. **/
