+++ OpenAIChat/src/main/kotlin/furhatos/app/openaichat/flow/main/idle.kt
@@ -1,0 +1,25 @@
+package furhatos.app.openaichat.flow
+
+import furhatos.app.openaichat.setting.activate
+import furhatos.app.openaichat.setting.hostPersona
+import furhatos.nlu.common.*
+import furhatos.flow.kotlin.*
+import furhatos.records.Location
+
+val Idle : State = state {
+    onEntry {
+        activate(hostPersona)
+        furhat.attendNobody()
+    }
+
+    onUserEnter {
+        furhat.attend(it)
+        goto(Greeting)
+    }
+
+}
+
+
+
+
+
