+++ OpenAIChat/src/main/kotlin/furhatos/app/openaichat/setting/interactionParams.kt
@@ -1,0 +1,4 @@
+package furhatos.app.openaichat.setting
+
+val maxNumberOfUsers = 2
+val distanceToEngage = 1.0
