+++ OpenAIChat/src/main/kotlin/furhatos/app/openaichat/utils/utils.kt
